backend/voice-agent/requirements.txt
# Pipecat Voice Pipeline Dependencies for ECS
# Pin versions for reproducibility

# Core Pipecat with Daily, Silero VAD, Deepgram, Cartesia, and SageMaker support
# - daily: WebRTC transport
# - silero: Voice Activity Detection
# - deepgram: Cloud STT API + SageMaker STT (DeepgramSageMakerSTTService)
# - cartesia: Cloud TTS API
# - aws: Bedrock LLM support
# - sagemaker: HTTP/2 BiDi streaming for SageMaker endpoints (requires Python >= 3.12)
pipecat-ai[daily,silero,deepgram,cartesia,aws,sagemaker]==0.0.102

# AWS SDK - let pip resolve version to match aiobotocore requirements
# aiobotocore 2.25.1 requires botocore>=1.40.46
# Do NOT pin boto3/botocore here - it conflicts with pipecat[aws]

# Audio processing
soundfile>=0.12.1
numpy>=1.26.0

# Async support
aiohttp>=3.9.0

# Logging
structlog>=24.1.0

# Environment and config
python-dotenv>=1.0.0

# A2A capability registry (lazy-imported, only used when enable_capability_registry=true)
# Provides A2AAgent client for fetching Agent Cards and invoking remote agents
strands-agents[a2a]>=1.27.0

# TTL cache for A2A response caching (latency optimization)
cachetools>=5.3.0

# Pipecat Flows - structured conversation framework for multi-agent handoff
# (lazy-imported, only used when enable_flow_agents=true)
# Provides FlowManager for in-process agent context switching with Bedrock support
pipecat-ai-flows>=0.0.23

# Note: Silero VAD uses ONNX Runtime (installed via pipecat-ai[silero] extra),
# not PyTorch. No torch/torchaudio dependency is needed.
